app/backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
Pillow==11.2.1
django-storages==1.14.4
boto3==1.38.11
exponent-server-sdk==2.1.0
django-cors-headers==4.9.0
coverage==7.13.5
requests==2.32.3  # used by scripts/download_fixture_images.py
